requirements.txt
baidu-aip==4.16.13
chardet==5.2.0
dotenv==0.9.9
Flask==3.1.0
openai==1.64.0
volcengine==1.0.174


python-dotenv~=1.0.1
requests~=2.32.3
websockets~=15.0.1